server: patch openssl with extra logging
This extra logging does the following:
  a. Confirms that we are triggering the expected codepaths in
  OpenSSL's code.
  b. Records the time taken for the decryption, measured from within
  OpenSSL.

Results from a. shows we are triggering the right codepath. An
interesting takeaway is that the codepath that it takes is stochastic
and dependent on the random bytes which the server generates.

I will run a new experiment tonight which compares the internal
measurements from b. with our external measurements. This should end
up being similar to the work in "Opportunities and Limits of Remote
Timing Attacks" by SCOTT A. CROSBY, DAN S. WALLACH and RUDOLF H. RIEDI
(http://www.cs.rice.edu/~dwallach/pub/crosby-timing2009.pdf).

Key questions:
  - Can we recover any of the expected behaviour using internal
  measurements? If not, this has no hope of working.
  - How accurate are our timings given s=1, s=7&n=400, s=20&n=1600,
  ...?