take network & internal measurements together then compare results
result shows that the network is not the source of issues in this
attack, theres something wrong with our setup which means we're not
triggering the bug?

REBASE COMMENT FROM THE FUTURE:
Deleted bit-samples.txt and internal-measurements.txt because they are too big. Hopefully you can regenerate them if you ever need to.