server: use gcc-2.96, redhat 9, apache-1.3.26, mod_ssl-2.8.10-1.3.26
Replaced ubuntu:latest on x86_64 with Redhat 9 i386 which is a much
closer vintage. This let me install and use their gcc-2.96
package. But I couldn't find a release of apache 1.3.27 for i386. So I
ended up picking apache-1.3.26, and the closest matching mod_ssl
version 2.8.10-1.3.26.

Maybe this is better, maybe it's worse. Will re-run experiments to
find out.