Revert gcc-2.96 and redhat 9 work
This reverts commits
     32a356986c32099419ff836a6a25781e3f407a83
     e28ca6fec29bb6db1e85453c7e85affc7f6d5674

which changed the servers runtime environment to Red Hat 9 instead of
Ubuntu rolling, and gcc-2.96. These versions gave setup closer to that
in the paper.

This doesn't seem to have helped, and adds a little complexity. So
I'm reverting these changes on master. I've kept those commits in a
branch "redhat-9-and-gcc-2.96" for prosperity.